simplified ankiconnect cd /home/r/anki-helper ; /usr/bin/env /home/r/anki-helper/.venv/bin/python /home/r/.vscode-server/extensions/ms-python.python-2023.22.1/pythonFiles/lib/python/debugpy/adapter/../../debugpy/launcher 51009 -- /home/r/anki-helper/tests/test_anki.py
