+++ config.py
@@ -21,10 +21,10 @@ OPENAI_API_KEY=os.getenv("OPENAI_API_KEY")
 # OPENAI_API_KEY="unsafe-secret-key-easy-to-steal"
 
 # Anki
-DEFAULT_DECKNAME = "anki-helper"
-DECKNAME = DEFAULT_DECKNAME
+ANKI_DEFAULT_DECKNAME = "anki-helper"
+ANKI_DECKNAME = ANKI_DEFAULT_DECKNAME
 
 # AnkiConnect
 ANKICONNECT_HOST="localhost"
 ANKICONNECT_PORT="8765"
-ANKICONNECT_ADDRESS = f"http://{ANKICONNECT_HOST}:{ANKICONNECT_PORT}"
+ANKICONNECT_URL = f"http://{ANKICONNECT_HOST}:{ANKICONNECT_PORT}"
